fitness_backend/accounts/api_urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register(r'profile', views.UserProfileViewSet, basename='profile')

urlpatterns = [
    path('register/', views.UserRegistrationView.as_view(), name='register'),
    path('login/', views.UserLoginView.as_view(), name='login'),
    path('logout/', views.logout_view, name='logout'),
    path('extended-profile/', views.ExtendedUserProfileView.as_view(), name='extended-profile'),
    path('settings/', views.UserSettingsView.as_view(), name='settings'),
    path('change-password/', views.change_password, name='change-password'),
    path('delete-account/', views.delete_account, name='delete-account'),
    path('search-users/', views.search_users, name='search-users'),
] + router.urls
