fitness_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
python-decouple==3.8
PyMySQL==1.1.0
celery==5.3.4
redis==5.0.1
django-allauth==0.57.0
dj-rest-auth==5.0.2
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
